fix groupid + add examples
