+++ bdd-pam-framework/pam-bdd-examples/project-to-test-with-bdd-parent/.vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "java.configuration.updateBuildConfiguration": "automatic"
+}
