+++ bdd-pam-framework/pam-bdd-examples/project-to-test-with-bdd-parent/pom.xml
@@ -1,0 +1,73 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://maven.apache.org/POM/4.0.0"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <parent>
+        <groupId>com.redhat.pam.bdd.examples</groupId>
+        <artifactId>pam-bdd-examples</artifactId>
+        <version>1.0</version>
+    </parent>
+
+    <artifactId>project-to-test-with-bdd-parent</artifactId>
+    <name>project-to-test-with-bdd-parent</name>
+    <packaging>pom</packaging>
+
+    <modules>
+        <module>project-to-test-with-bdd</module>
+        <module>project-to-test-with-bdd-test-data</module>
+        <module>project-to-test-with-bdd-wih</module>
+    </modules>
+
+    <properties>
+        <com.redhat.deployment.kjar.gav>${project.groupId}:project-to-test-with-bdd:${project.version}</com.redhat.deployment.kjar.gav>
+    </properties>
+
+    <dependencyManagement>
+        <dependencies>
+            <dependency>
+                <groupId>com.redhat.pam.bdd.examples</groupId>
+                <artifactId>project-to-test-with-bdd-test-data</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>com.redhat.pam.bdd.examples</groupId>
+                <artifactId>project-to-test-with-bdd-wih</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+        </dependencies>
+    </dependencyManagement>
+
+    <profiles>
+        <profile>
+            <id>junit-test</id>
+            <activation>
+                <activeByDefault>false</activeByDefault>
+            </activation>
+            <modules>
+                <module>project-to-test-with-bdd-junit-test</module>
+            </modules>
+            <build>
+                <plugins>
+                    <plugin>
+                        <groupId>org.apache.maven.plugins</groupId>
+                        <artifactId>maven-surefire-plugin</artifactId>
+                        <version>2.22.2</version>
+                        <configuration>
+                            <systemProperties>
+                                <property>
+                                    <name>kjar.deployment.gav</name>
+                                    <value>${com.redhat.deployment.kjar.gav}</value>
+                                </property>
+                                <property>
+                                    <name>kie.maven.settings.custom</name>
+                                    <value>/home/ippul/.m2/settings.xml</value>
+                                </property>
+                            </systemProperties>
+                        </configuration>
+                    </plugin>
+                </plugins>
+            </build>
+        </profile>
+    </profiles>
+</project>
