+++ bdd-pam-framework/pam-bdd-examples/project-to-test-with-bdd-parent/project-to-test-with-bdd-junit-test/src/test/resources/META-INF/services/org.redhat.pam.guice.BDDCustomContextProvider
@@ -1,0 +1,1 @@
+com.redhat.examples.guice.ContextProvider
