+++ bdd-pam-framework/pam-bdd-examples/project-to-test-with-bdd-parent/project-to-test-with-bdd-junit-test/src/test/resources/data/example-custom-object-com.redhat.examples.bdd.process.Document.json
