+++ bdd-pam-framework/pam-bdd-examples/project-to-test-with-bdd-parent/project-to-test-with-bdd-wih/src/main/java/com/redhat/examples/bdd/process/WIHDocument.java
@@ -1,0 +1,24 @@
+package com.redhat.examples.bdd.process;
+
+import org.jbpm.process.workitem.core.AbstractLogOrThrowWorkItemHandler;
+import org.kie.api.runtime.process.WorkItem;
+import org.kie.api.runtime.process.WorkItemManager;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.util.HashMap;
+
+public class WIHDocument extends AbstractLogOrThrowWorkItemHandler {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(WIHDocument.class);
+
+    @Override
+    public void executeWorkItem(WorkItem workItem, WorkItemManager manager) {
+        LOGGER.info("Work Item Handler class: {}", this.getClass());
+        manager.completeWorkItem(workItem.getId(), new HashMap<String, Object>());
+    }
+
+    @Override
+    public void abortWorkItem(WorkItem workItem, WorkItemManager workItemManager) {
+    }
+}
