+++ bdd-pam-framework/pam-bdd-examples/project-to-test-with-bdd-parent/project-to-test-with-bdd/global/defaultbusinessrulesicon.png
