+++ bdd-pam-framework/pam-bdd-examples/project-to-test-with-bdd-parent/project-to-test-with-bdd/src/main/resources/com/redhat/examples/bdd/process/process-with-custom-wih.bpmn
@@ -1,0 +1,104 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<bpmn2:definitions xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://www.omg.org/bpmn20" xmlns:bpmn2="http://www.omg.org/spec/BPMN/20100524/MODEL" xmlns:bpmndi="http://www.omg.org/spec/BPMN/20100524/DI" xmlns:bpsim="http://www.bpsim.org/schemas/1.0" xmlns:dc="http://www.omg.org/spec/DD/20100524/DC" xmlns:di="http://www.omg.org/spec/DD/20100524/DI" xmlns:drools="http://www.jboss.org/drools" id="_yoCVsGCKEeuZWbx6y7gEDg" xsi:schemaLocation="http://www.omg.org/spec/BPMN/20100524/MODEL BPMN20.xsd http://www.jboss.org/drools drools.xsd http://www.bpsim.org/schemas/1.0 bpsim.xsd http://www.omg.org/spec/DD/20100524/DC DC.xsd http://www.omg.org/spec/DD/20100524/DI DI.xsd " exporter="jBPM Process Modeler" exporterVersion="2.0" targetNamespace="http://www.omg.org/bpmn20">
+  <bpmn2:process id="project-to-test-with-bdd.process-with-custom-wih" drools:packageName="com.redhat.examples.bdd.process" drools:version="1.0" drools:adHoc="false" name="process-with-custom-wih" isExecutable="true" processType="Public">
+    <bpmn2:sequenceFlow id="_1B2885EA-910F-408D-A407-021DFCCBEA70" sourceRef="_9074FC5D-5AE7-48EA-88BF-0D1E03A47DB5" targetRef="_10ECBCB3-DED9-4157-9E5B-AD93EEE84EEE"/>
+    <bpmn2:sequenceFlow id="_2FAB5C3B-34F6-407B-92B7-4CF9498E2CB7" sourceRef="_10ECBCB3-DED9-4157-9E5B-AD93EEE84EEE" targetRef="_032DD9EE-9648-40E4-B18F-B826E525FE9A">
+      <bpmn2:extensionElements>
+        <drools:metaData name="isAutoConnection.source">
+          <drools:metaValue><![CDATA[true]]></drools:metaValue>
+        </drools:metaData>
+        <drools:metaData name="isAutoConnection.target">
+          <drools:metaValue><![CDATA[true]]></drools:metaValue>
+        </drools:metaData>
+      </bpmn2:extensionElements>
+    </bpmn2:sequenceFlow>
+    <bpmn2:task id="_10ECBCB3-DED9-4157-9E5B-AD93EEE84EEE" drools:taskName="CustomWorkItemHandler" name="CustomWorkItemHandler">
+      <bpmn2:extensionElements>
+        <drools:metaData name="elementname">
+          <drools:metaValue><![CDATA[CustomWorkItemHandler]]></drools:metaValue>
+        </drools:metaData>
+      </bpmn2:extensionElements>
+      <bpmn2:incoming>_1B2885EA-910F-408D-A407-021DFCCBEA70</bpmn2:incoming>
+      <bpmn2:outgoing>_2FAB5C3B-34F6-407B-92B7-4CF9498E2CB7</bpmn2:outgoing>
+      <bpmn2:ioSpecification id="_yoCVsWCKEeuZWbx6y7gEDg">
+        <bpmn2:dataInput id="_10ECBCB3-DED9-4157-9E5B-AD93EEE84EEE_TaskNameInputX" drools:dtype="Object" name="TaskName"/>
+        <bpmn2:inputSet id="_yoC8wGCKEeuZWbx6y7gEDg">
+          <bpmn2:dataInputRefs>_10ECBCB3-DED9-4157-9E5B-AD93EEE84EEE_TaskNameInputX</bpmn2:dataInputRefs>
+        </bpmn2:inputSet>
+        <bpmn2:outputSet id="_yoC8wWCKEeuZWbx6y7gEDg"/>
+      </bpmn2:ioSpecification>
+      <bpmn2:dataInputAssociation id="_yoC8wmCKEeuZWbx6y7gEDg">
+        <bpmn2:targetRef>_10ECBCB3-DED9-4157-9E5B-AD93EEE84EEE_TaskNameInputX</bpmn2:targetRef>
+        <bpmn2:assignment id="_yoC8w2CKEeuZWbx6y7gEDg">
+          <bpmn2:from xsi:type="bpmn2:tFormalExpression" id="_yoC8xGCKEeuZWbx6y7gEDg"><![CDATA[CustomWorkItemHandler]]></bpmn2:from>
+          <bpmn2:to xsi:type="bpmn2:tFormalExpression" id="_yoC8xWCKEeuZWbx6y7gEDg">_10ECBCB3-DED9-4157-9E5B-AD93EEE84EEE_TaskNameInputX</bpmn2:to>
+        </bpmn2:assignment>
+      </bpmn2:dataInputAssociation>
+    </bpmn2:task>
+    <bpmn2:startEvent id="_9074FC5D-5AE7-48EA-88BF-0D1E03A47DB5">
+      <bpmn2:outgoing>_1B2885EA-910F-408D-A407-021DFCCBEA70</bpmn2:outgoing>
+    </bpmn2:startEvent>
+    <bpmn2:endEvent id="_032DD9EE-9648-40E4-B18F-B826E525FE9A">
+      <bpmn2:incoming>_2FAB5C3B-34F6-407B-92B7-4CF9498E2CB7</bpmn2:incoming>
+    </bpmn2:endEvent>
+  </bpmn2:process>
+  <bpmndi:BPMNDiagram id="_yoC8xmCKEeuZWbx6y7gEDg">
+    <bpmndi:BPMNPlane id="_yoC8x2CKEeuZWbx6y7gEDg" bpmnElement="project-to-test-with-bdd.process-with-custom-wih">
+      <bpmndi:BPMNShape id="shape__032DD9EE-9648-40E4-B18F-B826E525FE9A" bpmnElement="_032DD9EE-9648-40E4-B18F-B826E525FE9A">
+        <dc:Bounds height="56.0" width="56.0" x="765.0" y="169.0"/>
+      </bpmndi:BPMNShape>
+      <bpmndi:BPMNShape id="shape__9074FC5D-5AE7-48EA-88BF-0D1E03A47DB5" bpmnElement="_9074FC5D-5AE7-48EA-88BF-0D1E03A47DB5">
+        <dc:Bounds height="56.0" width="56.0" x="341.0" y="169.0"/>
+      </bpmndi:BPMNShape>
+      <bpmndi:BPMNShape id="shape__10ECBCB3-DED9-4157-9E5B-AD93EEE84EEE" bpmnElement="_10ECBCB3-DED9-4157-9E5B-AD93EEE84EEE">
+        <dc:Bounds height="102.0" width="154.0" x="518.0" y="146.0"/>
+      </bpmndi:BPMNShape>
+      <bpmndi:BPMNEdge id="edge_shape__10ECBCB3-DED9-4157-9E5B-AD93EEE84EEE_to_shape__032DD9EE-9648-40E4-B18F-B826E525FE9A" bpmnElement="_2FAB5C3B-34F6-407B-92B7-4CF9498E2CB7">
+        <di:waypoint xsi:type="dc:Point" x="672.0" y="197.0"/>
+        <di:waypoint xsi:type="dc:Point" x="765.0" y="197.0"/>
+      </bpmndi:BPMNEdge>
+      <bpmndi:BPMNEdge id="edge_shape__9074FC5D-5AE7-48EA-88BF-0D1E03A47DB5_to_shape__10ECBCB3-DED9-4157-9E5B-AD93EEE84EEE" bpmnElement="_1B2885EA-910F-408D-A407-021DFCCBEA70">
+        <di:waypoint xsi:type="dc:Point" x="369.0" y="197.0"/>
+        <di:waypoint xsi:type="dc:Point" x="518.0" y="197.0"/>
+      </bpmndi:BPMNEdge>
+    </bpmndi:BPMNPlane>
+  </bpmndi:BPMNDiagram>
+  <bpmn2:relationship id="_yoC8yGCKEeuZWbx6y7gEDg" type="BPSimData">
+    <bpmn2:extensionElements>
+      <bpsim:BPSimData>
+        <bpsim:Scenario xsi:type="bpsim:Scenario" id="default" name="Simulationscenario">
+          <bpsim:ScenarioParameters xsi:type="bpsim:ScenarioParameters"/>
+          <bpsim:ElementParameters xsi:type="bpsim:ElementParameters" elementRef="_9074FC5D-5AE7-48EA-88BF-0D1E03A47DB5" id="_yoC8yWCKEeuZWbx6y7gEDg">
+            <bpsim:TimeParameters xsi:type="bpsim:TimeParameters">
+              <bpsim:ProcessingTime xsi:type="bpsim:Parameter">
+                <bpsim:NormalDistribution mean="0.0" standardDeviation="0.0"/>
+              </bpsim:ProcessingTime>
+            </bpsim:TimeParameters>
+          </bpsim:ElementParameters>
+          <bpsim:ElementParameters xsi:type="bpsim:ElementParameters" elementRef="_10ECBCB3-DED9-4157-9E5B-AD93EEE84EEE" id="_yoC8ymCKEeuZWbx6y7gEDg">
+            <bpsim:TimeParameters xsi:type="bpsim:TimeParameters">
+              <bpsim:ProcessingTime xsi:type="bpsim:Parameter">
+                <bpsim:NormalDistribution mean="0.0" standardDeviation="0.0"/>
+              </bpsim:ProcessingTime>
+            </bpsim:TimeParameters>
+            <bpsim:ResourceParameters xsi:type="bpsim:ResourceParameters">
+              <bpsim:Availability xsi:type="bpsim:Parameter">
+                <bpsim:FloatingParameter value="0.0"/>
+              </bpsim:Availability>
+              <bpsim:Quantity xsi:type="bpsim:Parameter">
+                <bpsim:FloatingParameter value="0.0"/>
+              </bpsim:Quantity>
+            </bpsim:ResourceParameters>
+            <bpsim:CostParameters xsi:type="bpsim:CostParameters">
+              <bpsim:UnitCost xsi:type="bpsim:Parameter">
+                <bpsim:FloatingParameter value="0.0"/>
+              </bpsim:UnitCost>
+            </bpsim:CostParameters>
+          </bpsim:ElementParameters>
+        </bpsim:Scenario>
+      </bpsim:BPSimData>
+    </bpmn2:extensionElements>
+    <bpmn2:source>_yoCVsGCKEeuZWbx6y7gEDg</bpmn2:source>
+    <bpmn2:target>_yoCVsGCKEeuZWbx6y7gEDg</bpmn2:target>
+  </bpmn2:relationship>
+</bpmn2:definitions>
