+++ bdd-pam-framework/pam-bdd-parent/pom.xml
@@ -3,7 +3,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.redhat</groupId>
+    <groupId>com.redhat.pam.bdd</groupId>
     <artifactId>pam-bdd-parent</artifactId>
     <packaging>pom</packaging>
     <name>pam-bdd-parent</name>
@@ -59,7 +59,7 @@
                 <version>${io.cucumber.version}</version>
             </dependency>
             <dependency>
-                <groupId>com.redhat</groupId>
+                <groupId>com.redhat.pam.bdd</groupId>
                 <artifactId>pam-bdd-junit-commons</artifactId>
                 <version>${project.version}</version>
             </dependency>
