--- bdd-pam-framework/project-to-test-with-bdd-parent/project-to-test-with-bdd-junit-test/src/test/resources/META-INF/services/org.ippul.pam.guice.BDDCustomContextProvider
@@ -1,1 +1,0 @@
-com.ippul.examples.guice.ContextProvider
