--- bdd-pam-framework/project-to-test-with-bdd-parent/project-to-test-with-bdd-junit-test/src/test/resources/logback.xml
@@ -1,22 +1,0 @@
-<configuration>
-
-    <appender name="CONSOLE" class="ch.qos.logback.core.ConsoleAppender">
-        <layout class="ch.qos.logback.classic.PatternLayout">
-            <Pattern>
-                %d{HH:mm:ss.SSS} [%t] %-5level %logger{36} - %msg%n
-            </Pattern>
-        </layout>
-    </appender>
-
-    <logger name="org.ippul" level="debug" additivity="false">
-        <appender-ref ref="CONSOLE"/>
-    </logger>
-    <logger name="org.hibernate" level="info" additivity="false">
-        <appender-ref ref="CONSOLE"/>
-    </logger>
-
-    <root level="debug">
-        <appender-ref ref="CONSOLE"/>
-    </root>
-
-</configuration>
