--- bdd-pam-framework/project-to-test-with-bdd-parent/project-to-test-with-bdd-test-data/pom.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://maven.apache.org/POM/4.0.0"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>org.ippul.pam.bdd.examples</groupId>
-        <artifactId>project-to-test-with-bdd-parent</artifactId>
-        <version>1.0</version>
-    </parent>
-
-    <artifactId>project-to-test-with-bdd-test-data</artifactId>
-    <packaging>jar</packaging>
-    <name>project-to-test-with-bdd-test-data</name>
-
-</project>
