--- bdd-pam-framework/project-to-test-with-bdd-parent/project-to-test-with-bdd/pom.xml
@@ -1,75 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
-         xmlns="http://maven.apache.org/POM/4.0.0">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>org.ippul.pam.bdd.examples</groupId>
-        <artifactId>project-to-test-with-bdd-parent</artifactId>
-        <version>1.0</version>
-    </parent>
-
-    <artifactId>project-to-test-with-bdd</artifactId>
-    <packaging>kjar</packaging>
-    <name>project-to-test-with-bdd</name>
-    <description></description>
-    <dependencies>
-        <dependency>
-            <groupId>org.ippul.pam.bdd.examples</groupId>
-            <artifactId>project-to-test-with-bdd-test-data</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.ippul.pam.bdd.examples</groupId>
-            <artifactId>project-to-test-with-bdd-wih</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.kie</groupId>
-            <artifactId>kie-api</artifactId>
-            <version>7.44.0.Final-redhat-00003</version>
-            <scope>provided</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.kie</groupId>
-            <artifactId>kie-internal</artifactId>
-            <version>7.44.0.Final-redhat-00003</version>
-            <scope>provided</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.optaplanner</groupId>
-            <artifactId>optaplanner-core</artifactId>
-            <version>7.44.0.Final-redhat-00003</version>
-            <scope>provided</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.optaplanner</groupId>
-            <artifactId>optaplanner-persistence-jaxb</artifactId>
-            <version>7.44.0.Final-redhat-00003</version>
-            <scope>provided</scope>
-        </dependency>
-        <dependency>
-            <groupId>junit</groupId>
-            <artifactId>junit</artifactId>
-            <version>4.12</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>com.thoughtworks.xstream</groupId>
-            <artifactId>xstream</artifactId>
-            <version>1.4.11.1-redhat-00001</version>
-            <scope>test</scope>
-        </dependency>
-    </dependencies>
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.kie</groupId>
-                <artifactId>kie-maven-plugin</artifactId>
-                <version>7.44.0.Final-redhat-00003</version>
-                <extensions>true</extensions>
-            </plugin>
-        </plugins>
-    </build>
-</project>
