--- bdd-pam-framework/project-to-test-with-bdd-parent/project-to-test-with-bdd/src/main/resources/META-INF/kie-deployment-descriptor.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<deployment-descriptor xsi:schemaLocation="http://www.jboss.org/jbpm deployment-descriptor.xsd" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-    <persistence-unit>org.jbpm.domain</persistence-unit>
-    <audit-persistence-unit>org.jbpm.domain</audit-persistence-unit>
-    <audit-mode>JPA</audit-mode>
-    <persistence-mode>JPA</persistence-mode>
-    <runtime-strategy>PER_PROCESS_INSTANCE</runtime-strategy>
-    <marshalling-strategies/>
-    <event-listeners/>
-    <task-event-listeners/>
-    <globals/>
-    <work-item-handlers>
-        <work-item-handler>
-            <resolver>mvel</resolver>
-            <identifier>new com.ippul.examples.bdd.process.WIHDocument()</identifier>
-            <parameters/>
-            <name>CustomWorkItemHandler</name>
-        </work-item-handler>
-    </work-item-handlers>
-    <environment-entries/>
-    <configurations/>
-    <required-roles/>
-    <remoteable-classes>
-        <remoteable-class>com.ippul.examples.bdd.process.Document</remoteable-class>
-    </remoteable-classes>
-    <limit-serialization-classes>true</limit-serialization-classes>
-</deployment-descriptor>
